requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
scikit-learn
requests
